Ingres migration, integer money storage, and display format change

+++ SpringMvcTest/db/create_products.sql
@@ -2,6 +2,6 @@ DROP TABLE products;
 CREATE TABLE products (
   id INTEGER NOT NULL PRIMARY KEY,
   description varchar(255),
-  price decimal(15,2)
+  price integer 
 );
 CREATE INDEX products_description ON products(description);
